Bug fixes und Aurfäumen
Sauberer Code für die Speichern funktion und paar exceptions gefangen. Nicht perfekt aber keine OutOfBounce mehr 
Layout fast final
BackButton
NeuKunden Anlegen funktion

+++ src/main/java/de/fhbielefeld/pmt/JPAEntities/Costs.java
@@ -17,7 +17,7 @@ public class Costs implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int costsID;
+	private long costsID;
 	private String costType;
 	private String description;
 	private double incurredCosts;
@@ -51,7 +51,7 @@ public class Costs implements Serializable {
 	 * @param 
 	 */
 	// Set-Methode nicht vorhanden, soll nicht veraendert werden 
-	public int getCostsID() {
+	public long getCostsID() {
 		return costsID;
 	}
